refactor!: merge serializer project
