--- src/KafkaFlow.BatchConsume/AssemblyInfo.cs
@@ -1,4 +1,0 @@
-using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("KafkaFlow.UnitTests")]
-[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
