--- src/KafkaFlow.BatchConsume/KafkaFlow.BatchConsume.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-        <PackageId>KafkaFlow.BatchConsume</PackageId>
-        <Description>A KafkaFlow middleware to consume messages in batches</Description>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\KafkaFlow.Abstractions\KafkaFlow.Abstractions.csproj" />
-      <ProjectReference Include="..\KafkaFlow\KafkaFlow.csproj" />
-    </ItemGroup>
-
-</Project>
