+++ src/KafkaFlow.Serializer.SchemaRegistry.ConfluentJson/KafkaFlow.Serializer.SchemaRegistry.ConfluentJson.csproj
@@ -15,7 +15,6 @@
     <ItemGroup>
         <ProjectReference Include="..\KafkaFlow.Abstractions\KafkaFlow.Abstractions.csproj" />
         <ProjectReference Include="..\KafkaFlow.SchemaRegistry\KafkaFlow.SchemaRegistry.csproj" />
-        <ProjectReference Include="..\KafkaFlow.Serializer\KafkaFlow.Serializer.csproj" />
     </ItemGroup>
 
 </Project>
