+++ src/KafkaFlow.Serializer.SchemaRegistry.ConfluentJson/ProducerConfigurationBuilderExtensions.cs
@@ -2,6 +2,7 @@
 {
     using Confluent.SchemaRegistry.Serdes;
     using KafkaFlow.Configuration;
+    using KafkaFlow.Serializer;
     using KafkaFlow.Serializer.SchemaRegistry;
 
     /// <summary>
