+++ src/KafkaFlow.Serializer.SchemaRegistry.ConfluentProtobuf/KafkaFlow.Serializer.SchemaRegistry.ConfluentProtobuf.csproj
@@ -16,6 +16,5 @@
     <ItemGroup>
         <ProjectReference Include="..\KafkaFlow.Abstractions\KafkaFlow.Abstractions.csproj" />
         <ProjectReference Include="..\KafkaFlow.SchemaRegistry\KafkaFlow.SchemaRegistry.csproj" />
-        <ProjectReference Include="..\KafkaFlow.Serializer\KafkaFlow.Serializer.csproj" />
     </ItemGroup>
 </Project>
