+++ src/KafkaFlow.UnitTests/KafkaFlow.UnitTests.csproj
@@ -31,11 +31,8 @@
     <ItemGroup>
         <ProjectReference Include="..\KafkaFlow.Admin.WebApi\KafkaFlow.Admin.WebApi.csproj" />
         <ProjectReference Include="..\KafkaFlow.Admin\KafkaFlow.Admin.csproj" />
-        <ProjectReference Include="..\KafkaFlow.BatchConsume\KafkaFlow.BatchConsume.csproj" />
-        <ProjectReference Include="..\KafkaFlow.Compressor\KafkaFlow.Compressor.csproj" />
         <ProjectReference Include="..\KafkaFlow.LogHandler.Microsoft\KafkaFlow.LogHandler.Microsoft.csproj" />
         <ProjectReference Include="..\KafkaFlow.Serializer.JsonCore\KafkaFlow.Serializer.JsonCore.csproj" />
-        <ProjectReference Include="..\KafkaFlow.Serializer\KafkaFlow.Serializer.csproj" />
         <ProjectReference Include="..\KafkaFlow.Serializer.NewtonsoftJson\KafkaFlow.Serializer.NewtonsoftJson.csproj" />
         <ProjectReference Include="..\KafkaFlow.TypedHandler\KafkaFlow.TypedHandler.csproj" />
         <ProjectReference Include="..\KafkaFlow\KafkaFlow.csproj" />
