+++ src/KafkaFlow/Compressor/CompressorConsumerMiddleware.cs
