+++ src/KafkaFlow/Compressor/ConfigurationBuilderExtensions.cs
