+++ src/KafkaFlow/Serializer/ProducerMiddlewareConfigurationBuilder.cs
@@ -1,4 +1,4 @@
-﻿namespace KafkaFlow
+﻿namespace KafkaFlow.Serializer
 {
     using System;
     using KafkaFlow.Configuration;
